requirements.txt
black==22.3.0
click>=7.1
flake8==4.0.1
loguru
numpy
opencv-python==4.5.5.62
pyyaml>=6.0.0